backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "securities-trading-assistant"
version = "1.0.0"
description = "证券交易助手Agent平台 - Securities Trading Assistant Agent Platform"
requires-python = ">=3.10"
dependencies = [
    "strands-agents>=1.13.0",
    "strands-agents-tools>=0.1.0",
    "bedrock-agentcore>=1.0.3",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "redis>=5.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "boto3>=1.34.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.0",
    "bcrypt>=4.0.0,<5.0.0",
    "python-multipart>=0.0.6",
    "aiohttp>=3.9.0",
    "apscheduler>=3.10.0",
    "jinja2>=3.1.0",
    "nest-asyncio>=1.6.0",
    "playwright>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[tool.setuptools.packages.find]
where = ["."]
